pyproject.toml
[project]
name = "swan-quote-generator"
version = "1.0.0"
description = "AI-powered quote generator using FastAPI and Google Gemini"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "mangum==0.17.0",
    "pydantic==2.8.2",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "google-generativeai==0.8.3",
    "python-multipart==0.0.6",
    "tenacity==8.2.3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
# Set the maximum line length
line-length = 100

# Target Python 3.11+
target-version = "py311"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "static",
    "node_modules",
    ".pytest_cache",
    ".ruff_cache",
]

[tool.ruff.lint]
# Enable specific linting rules
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
]

# Ignore specific rules
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function calls in argument defaults (needed for FastAPI)
    "N805",   # First argument of a method should be named `self` (pydantic validators)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
# Ignore import rules in __init__.py files
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
# Configure import sorting
known-first-party = ["app"]
lines-after-imports = 2

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with 4 spaces
indent-style = "space"

# Use Unix line endings
line-ending = "auto"
